Fix up md5 impl and some benchmark evalsgf cmds

+++ cpp/benchmark.cpp
@@ -33,7 +33,7 @@ int MainCmds::benchmark(int argc, const char* const* argv) {
     TCLAP::ValueArg<string> modelFileArg("","model","Neural net model file to use",true,string(),"FILE");
     TCLAP::ValueArg<string> sgfFileArg("","sgf", "Optional game to sample positions from (default: uses a built-in-set of positions)",false,string(),"FILE");
     TCLAP::ValueArg<int> boardSizeArg("","boardsize", "Size of board to benchmark on (9-19), default 19",false,-1,"SIZE");
-    TCLAP::ValueArg<int64_t> visitsArg("v","visits","How many visits to use per search (default " + Global::int64ToString(defaultMaxVisits) + ")",false,defaultMaxVisits,"VISITS");
+    TCLAP::ValueArg<long> visitsArg("v","visits","How many visits to use per search (default " + Global::int64ToString(defaultMaxVisits) + ")",false,(long)defaultMaxVisits,"VISITS");
     TCLAP::ValueArg<string> threadsArg("t","threads","Test using these many threads, comma-separated (default 1,2,4,6,8,12,16)",false,string("1,2,4,6,8,12,16"),"THREADS");
     TCLAP::ValueArg<int> numPositionsPerGameArg("n","numpositions","How many positions to sample from a game (default 10)",false,10,"NUM");
     cmd.add(configFileArg);
@@ -48,7 +48,7 @@ int MainCmds::benchmark(int argc, const char* const* argv) {
     modelFile = modelFileArg.getValue();
     sgfFile = sgfFileArg.getValue();
     boardSize = boardSizeArg.getValue();
-    maxVisits = visitsArg.getValue();
+    maxVisits = (int64_t)visitsArg.getValue();
     desiredThreadsStr = threadsArg.getValue();
     numPositionsPerGame = numPositionsPerGameArg.getValue();
   }
